api/requirements.txt
absl-py==2.3.1
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
ale-py==0.11.2
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
async-timeout==5.0.1
attrs==25.4.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
babel==2.17.0
backcall==0.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
bleach==6.3.0
blinker==1.9.0
boltons==21.0.0
bracex==2.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
click-option-group==0.5.9
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.3
-e git+https://github.com/cage-challenge/CybORG.git@2742b5e0ce4330c9b14006b38acd3b5ebe00d6fd#egg=CybORG
cycler==0.12.1
datasets==4.8.4
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
dm-tree==0.1.9
docker==7.1.0
docopt==0.6.2
docutils==0.22.4
etils==1.13.0
exceptiongroup==1.2.2
executing==2.2.1
face==26.0.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.1.0
glfw==2.10.0
glom==25.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gym==0.23.1
gym-notices==0.1.0
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.11.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
invoke==2.2.1
ipykernel==7.1.0
ipython==8.12.3
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
lz4==4.4.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.23.3
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.4.0
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.2
notebook_shim==0.2.4
numpy==1.26.4
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.5
peewee==3.19.0
peft==0.19.1
pettingzoo==1.24.3
pickleshare==0.7.5
pillow==12.1.0
pipreqs==0.5.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.1
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.11.0
PyNaCl==1.6.2
PyOpenGL==3.1.10
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.9.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.5.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.14
safetensors==0.7.0
sb3_contrib==2.7.1
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
semgrep==1.153.1
Send2Trash==2.1.0
shellingham==1.5.4
Shimmy==0.2.1
six==1.17.0
soupsieve==2.8.3
sqlmap==1.10.2
sse-starlette==3.3.2
stable_baselines3==2.7.1
stack-data==0.6.3
starlette==0.52.1
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
terminado==0.18.1
tifffile==2025.5.10
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.0.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==5.5.4
trl==1.2.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.11.3
uvicorn==0.41.0
wcmatch==8.5.2
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==1.17.3
xxhash==3.6.0
yarg==0.1.9
yarl==1.23.0
zaproxy==0.5.0
zipp==3.23.0
